feat: Authenticate user in Save, Update & Delete routes

+++ routes/post.js
@@ -1,7 +1,8 @@
 const express = require("express");
 const router = express.Router();
 const Post = require("../models/post");
-const User = require("../models/user");
+const _ = require("lodash");
+const auth = require("../middleware/auth");
 
 // View all post
 router.get("/", (req, res) => {
@@ -32,37 +33,46 @@ router.get("/author/:authorId", (req, res) => {
 });
 
 // Save new post
-router.post("/", (req, res) => {
-  User.findById(req.body.author).then(response => {
-    if (response) {
-      Post.create(req.body)
-        .then(response => {
-          const { _id, author, message, date } = response;
-          res.send({ _id, author, message, date });
-        })
-        .catch(err => res.status(400).send(err.message));
-    } else {
-      res.status(400).send("Bad Request: Invalid User");
-    }
-  });
+router.post("/", auth, (req, res) => {
+  req.body.author = req.user._id;
+  Post.create(_.pick(req.body, ["author", "message", "date"]))
+    .then(response =>
+      res.send(_.pick(response, ["_id", "author", "message", "date"]))
+    )
+    .catch(err => res.status(400).send(err.message));
 });
 
 // Update post
-router.put("/:id", (req, res) => {
-  Post.updateOne({ _id: req.params.id }, { $set: req.body })
+router.put("/:id", auth, async (req, res) => {
+  const post = await Post.findOne({ _id: req.params.id });
+  req.body.date = Date.now();
+
+  if (req.user._id !== post.author.toString())
+    return res.status(401).send("Permission denied.");
+  Post.updateOne(
+    { _id: req.params.id },
+    { $set: _.pick(req.body, ["date", "message"]) }
+  )
     .then(response => res.send(response))
     .catch(err => res.status(400).send(err.message));
 });
 
 // Delete post
-router.delete("/:id", (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
+  const post = await Post.findOne({ _id: req.params.id });
+  if (req.user._id !== post.author.toString())
+    return res.status(401).send("Permission denied.");
+
   Post.deleteOne({ _id: req.params.id })
     .then(response => res.send(response))
     .catch(err => res.status(400).send(err.message));
 });
 
 // Delete all post by author
-router.delete("/author/:authorId", (req, res) => {
+router.delete("/author/:authorId", auth, (req, res) => {
+  if (req.user._id !== req.params.authorId)
+    return res.status(401).send("Permission denied.");
+
   Post.deleteMany({ author: req.params.authorId })
     .then(response => res.send(response))
     .catch(err => res.status(400).send(err.message));
